Created Backend File Management and started on UI
Created backend file management, as well as started on the UI for the program. Program Handler and Program Engine will be incorporated later as I need to first create the ability to make and retrieve flashcards before I can make a game.

Code needs commenting, better explanation, and can be optimized, however that will be taken care of in the future.